+++ vagabond/core/CubicGrid.h
@@ -30,7 +30,7 @@ public:
 	void setRealDim(float dim);
 	void setRecipDim(float dim);
 
-	virtual double resolution(int i, int j, int k);
+	virtual float resolution(int i, int j, int k);
 	virtual void real2Voxel(glm::vec3 &real);
 	virtual glm::vec3 reciprocal(int h, int k, int l);
 private:
